update tests after rebase
